+++ src/main/java/com/codessquad/qna/domain/AbstractEntity.java
@@ -1,16 +1,19 @@
 package com.codessquad.qna.domain;
 
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
+import javax.persistence.*;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
 @MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
 public class AbstractEntity {
+    private static final DateTimeFormatter FORMATTER_PATTERN = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -18,5 +21,26 @@ public class AbstractEntity {
 
     @CreatedDate
     private LocalDateTime createdDateTime;
+
+    @LastModifiedDate
+    private LocalDateTime modifiedDateTime;
+
+    public Long getId() {
+        return id;
+    }
+
+    public boolean matchId(Long newId) {
+        return this.id.equals(newId);
+    }
+
+    @JsonProperty("createdDateTime")
+    public String getCreatedDateTime() {
+        return createdDateTime.format(FORMATTER_PATTERN);
+    }
+
+    @JsonProperty("modifiedDateTime")
+    public String getModifiedDateTime() {
+        return modifiedDateTime.format(FORMATTER_PATTERN);
+    }
 
 }
