+++ src/main/java/com/codessquad/qna/domain/Question.java
@@ -9,12 +9,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 @Entity
-public class Question {
-    private static final DateTimeFormatter FORMATTER_PATTERN = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+public class Question extends AbstractEntity{
 
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_question_writer"))
@@ -27,8 +22,6 @@ public class Question {
     @Column(nullable = false, length = 2000)
     @NotBlank(message = "내용은 필수 입력 값입니다.")
     private String contents;
-
-    private LocalDateTime createdDateTime = LocalDateTime.now();
 
     @OneToMany(mappedBy = "question")
     @JsonManagedReference
@@ -52,7 +45,6 @@ public class Question {
     public void updateQuestion(String title, String contents) {
         this.title = title;
         this.contents = contents;
-        this.createdDateTime = LocalDateTime.now();
     }
 
     public void setWriter(User writer) {
@@ -67,14 +59,6 @@ public class Question {
         this.contents = contents;
     }
 
-    public void setCreatedDateTime(LocalDateTime createdDateTime) {
-        this.createdDateTime = createdDateTime;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
     public User getWriter() {
         return writer;
     }
@@ -85,10 +69,6 @@ public class Question {
 
     public String getContents() {
         return contents;
-    }
-
-    public String getCreatedDateTime() {
-        return createdDateTime.format(FORMATTER_PATTERN);
     }
 
     public List<Answer> getAnswers() {
@@ -130,11 +110,11 @@ public class Question {
     @Override
     public String toString() {
         return "Question{" +
-                "id=" + id +
+                "id=" + getId() +
                 ", writer=" + writer +
                 ", title='" + title + '\'' +
                 ", contents='" + contents + '\'' +
-                ", createdDateTime=" + createdDateTime +
+                ", createdDateTime=" + getCreatedDateTime() +
                 '}';
     }
