+++ src/main/resources/import.sql
@@ -1,7 +1,7 @@
-INSERT INTO USER (USER_ID, PASSWORD, NAME, EMAIL) VALUES ('bat','aaaa','kim','bat@gmial.com');
-INSERT INTO USER (USER_ID, PASSWORD, NAME, EMAIL) VALUES ('iron','bbbb','lee','iron@gmail.com');
+INSERT INTO USER (USER_ID, PASSWORD, NAME, EMAIL, CREATED_DATE_TIME) VALUES ('bat','aaaa','kim','bat@gmial.com','2021-03-20 15:11');
+INSERT INTO USER (USER_ID, PASSWORD, NAME, EMAIL, CREATED_DATE_TIME) VALUES ('iron','bbbb','lee','iron@gmail.com','2021-03-21 15:11');
 
-INSERT INTO QUESTION (CONTENTS, CREATED_DATE_TIME,  IS_DELETE, TITLE, WRITER_ID) VALUES ('테스트용입니다.','2021-03-20 15:11', 'FALSE', 'Test', 1);
+INSERT INTO QUESTION (CONTENTS, CREATED_DATE_TIME, MODIFIED_DATE_TIME,  IS_DELETE, TITLE, WRITER_ID) VALUES ('테스트용입니다.','2021-03-20 15:11', '2021-03-21 18:11', 'FALSE', 'Test', 1);
 INSERT INTO ANSWER (CONTENTS, CREATED_DATE_TIME, IS_DELETE, QUESTION_ID, WRITER_ID) VALUES ('answer','2021-03-20 16:11', 'TRUE', 1, 1);
 INSERT INTO ANSWER (CONTENTS, CREATED_DATE_TIME, IS_DELETE, QUESTION_ID, WRITER_ID) VALUES ('answer','2021-03-20 17:11', 'FALSE', 1, 1);
